pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling", "hatch-vcs"]

[project]
name = "cscd"
dynamic = ["version"]
description = "**cscd** (pronounced cascade) is a content-addressed workflow orchestration tool that brings the power of content-addressable storage (CAS) to everyday development workflows. Companion to cascache."
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "ladidadida", email = "stefan@dalada.de" },
]
requires-python = ">=3.13"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "typer>=0.12.0",
  "pyyaml>=6.0.0",
  "pydantic>=2.0.0",
  "networkx>=3.0",
  "rich>=13.0.0",
  "aiofiles>=23.0.0",
  "anyio>=4.0.0",
  "grpcio>=1.64.0",
  "protobuf>=5.26.0",
]

[project.urls]
Homepage = "https://gitlab.com/cascascade/cscd"
Repository = "https://gitlab.com/cascascade/cscd"
Issues = "https://gitlab.com/cascascade/cscd/-/issues"
Documentation = "https://gitlab.com/cascascade/cscd/-/blob/main/README.md"

[dependency-groups]
dev = [
  "pytest>=9.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=7.0.0",
  "types-networkx>=3.5.0",
  "types-pyyaml>=6.0.12",
  "ruff>=0.3.0",
  "pyright>=1.1.0",
  "build>=1.0",
  "twine>=4.0",
  "grpcio-tools>=1.64.0",
]
# Optional: cascache integration tests (requires Docker or local cascache)
# Install with: uv sync --group cascache
cascache = [
  "cascache @ git+https://gitlab.com/cascascade/cascache@main",
]

[project.scripts]
cscd = "cscd.cli:main"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.sdist]
include = [
  "src",
  "tests",
  "README.md",
  "LICENSE",
]

[tool.pytest.ini_options]
addopts = "--strict-markers"
markers = [
  "component: marks tests as component-level (integration, CLI, fixtures)",
  "cascache: marks tests requiring real cascache server (integration-cascache)",
]
testpaths = ["tests"]

[tool.pyright]
typeCheckingMode = "basic"
pythonVersion = "3.13"
include = ["src", "tests"]
exclude = [
  "**/node_modules",
  "**/__pycache__",
  ".venv",
  "**/generated",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
  "E",    # pycodestyle errors
  "F",    # pyflakes
  "I",    # isort
  "W",    # pycodestyle warnings
  "C90",  # mccabe
  "UP",   # pyupgrade
]
extend-ignore = [
  "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
